Refactored link creation
